refactor: decouple wechat_slim into engine modules (scanner, cleaner, dedup, web, common)
